Cart. Mostly working. Issues in description below.
Flag lines 136, 185, 232, 278, 324, and 371.
Everything but line 371 is the same issues, which is a nullpointer exception when attempting to setManaged(false) the parent Hbox for the individual item components. In theory that should prevent them from attempting to generate at all if they aren't needed, but I can't get that to work.

371 is just a null pointer exception that I don't understand when attempting to implement the subtotal.

The only other issue is that when you remove the last instance of an item from the order, the way this is written it needs to remove that corresponding element from the uniqueItems arraylist that I made. Don't know how to do that without breaking it.

Like the menu this isn't really extensible to arbitrary size yet, that's something that'll have to be worked on in future.